+++ backend/src/main/java/backend/hanpum/domain/auth/dto/requestDto/SendEmailAuthCodeReqDto.java
@@ -1,4 +1,4 @@
-package backend.hanpum.domain.auth.dto;
+package backend.hanpum.domain.auth.dto.requestDto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
@@ -6,16 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
 
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CheckEmailAuthCodeReqDto {
+public class SendEmailAuthCodeReqDto {
     @NotNull
     @Email
     String email;
-    @NotNull
-    String inputAuthCode;
 }
